Switched from AppDomain to process spawning
The main process only watches the source folder. When a change is
detected, it spawns a child process that compiles and hosts Nancy with
the compiled binaries. Existing spawned processes are killed before
that. All output from the child processes is redirected to the main
window.

In my opinion, spawned processes are a lot easier to deal with than
custom AppDomains. I was able to fix nagging issues like changing the
current directory very easily for example.

Also, I added a check to only load the same assembly, based on its file
name, once.